Initial Commit
Git was initialised on top of a copy of an existing Chassis vagrantbox I
was using for demonstration purposes. I have modified the configuration
a little and moved the configuration file into the project root so it is
included in the repo. Chassis itself is loaded as a submodule to keep
the filesize to a minimum. The theme folder is mounted from the repo
root into the server via Virtualbox shared folder I have added to
`config.yaml`.