Adjust spacing between about bio and image
Pierce wanted this to be the same as the content below but this doesn't
look right, so I've gone for half this size. Visually this makes it look
like there is equal space between menu, bio, and image. Creates a visual
break between the header and CV sections which I think works well.