Create external links on about page
Added a conditional to the about page template which looks for a custom
field called 'link', and adds the link to the CV item title if it finds
one. I've added the link this way to ignore the main content, so in the
future the site can be changed to link to an internal page if site
content is present.