Fix gulpfile bugs
The `php` task was copying directories across while respecting the
directory stucture. WordPress requires a flat folder structure, so I
added `gulp-rename` as a dev-dependency and used this to flatten the
directory on export.

There were some duplicated slashes in the blobs for the `js` task which
weren't causing issues but which I fixed anyway.

The `img` task was including the asset files (not necessary for ./dist)
and also copying the screenshot file which was being copied to the root
dist folder in a separate task, so I created new blobs to ignore these.

It turns out that you can't create negating globs when also inlcuding a
property from another object, which is what I'd been doing in using
dir.src and dir.build essentially as aliases. This defeats the point of
using the dir object, so at some point I'll refactor the gulpfile and
bake the directories directly into the tasks, which should also decrease
the lines on the page significantly.