Update sidescroll gallery function and style
The sidescrolling gallery is now a collection of figure elements in
div.gallery-sidescroll, styled as flex with nowrap to keep the figure
elements in one row, with overflow-x: scroll on the parent element. The
first and last elements align with the content using calc(), and the
original gallery shortcode is stripped from the content. The .main
element on these pages is now full-width to allow the gallery to scroll
off the side of the page, and the elements which need to have a fixed
width are styled separately. I have commented these changes fairly
liberally in the scss.