Hide location separator on About loop when missing
About page previously had a dash after the title. However, this caused
the dash to appear floating there when no location is supplied, which is
a very forseeable situation seeing as the location is hidden behind a
custom field. The dash is generated by CSS (it doesn't appear on mobile
views), so I moved it from an ::after element on the title to a ::before
element on the location itself.