Add cv post pages logic and styles
If an entry on the About page has content but no associated URL in the
custom fields, it will instead link to an internal post. This means that
Pierce can duplicate interviews and articles on the site whilst the
articles are still active, and if they go offline he can simply remove
the link and the content is replicated on his own site.

The stylesheets for these pages are very similar to project pages, so
I've refactored the `project-` qualifier within the style classes to
simply read `single`, and added one extra rule to left-align the content
within the new linked articles.