Add space below portfolio galleries
There was no space between galleries and the date below. This wasn't
obvious on my local test because the first image was shorter than the
others, and they are top-aligned.

Other portfolio pages have an image at the top and the title below. This
is due to the top image being a featured image, rather than a gallery.
The inconsistency is annoying, but it will likely require some sort of
grep operation to fix and will need to take into account cases where
there is more than one gallery on a page.

For the time being a workable solution is to just add some margin to the
bottom of a gallery to match the margins on portfolio headers.